fix bug that prevents RegMon from working on mips platforms & ath9k
This patch changes where the hrtimer_start() and hrtimer_cancel()
functions are called in the ath9 code base. From now on we prevent
register reading while the WiFi hw is in reset state, which seems to cause
of the interrupt hang before.
RegMons proper register reading is tested on an TPLInk tl-wr1043nd v3.

A big thanks to Felix Fietkau (aka nbd) from lede-project for his help to
solve this bug!

Signed-off-by: Thomas Huehn <thomas@net.t-labs.tu-berlin.de>